Beginning of charts --> To really improved !!

+++ data/src/main/java/thomas/example/com/data/entity/ScoreEntity.kt
@@ -1,6 +1,6 @@
 package thomas.example.com.data.entity
 
 data class ScoreEntity(var idScore: String = "",
-                 var valueScore: Int = 0,
+                 var valueScore: Float = 0f,
                  var dateScore: String = "",
                  var idSong: String = "")
