+++ data/src/main/java/thomas/example/com/data/entity/remote/score/ScoreRemoteEntity.kt
@@ -3,6 +3,6 @@ package thomas.example.com.data.entity.remote.score
 import com.google.gson.annotations.SerializedName
 
 data class ScoreRemoteEntity(@SerializedName("idScore") var idScore: String = "",
-                             @SerializedName("valueScore") var valueScore: Int = 0,
+                             @SerializedName("valueScore") var valueScore: Float = 0f,
                              @SerializedName("dateScore") var dateScore: String = "",
                              @SerializedName("idSong") var idSong: String = "")
