+++ domain/src/main/java/thomas/example/com/model/Score.kt
@@ -1,6 +1,6 @@
 package thomas.example.com.model
 
 data class Score(var idScore: String = "",
-                   var valueScore: Int = 0,
+                   var valueScore: Float = 0f,
                    var dateScore: String = "",
                    var idSong: String = "")
